requirements.txt
python-telegram-bot>=20.7
fastapi>=0.115.0
uvicorn>=0.30.0
apscheduler>=3.10.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
httpx>=0.27.0
